Deprecate -O,--optimize and Configuration setOptimizeDatabase()

+++ src/org/opensolaris/opengrok/index/IndexDatabase.java
@@ -53,8 +53,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.zip.GZIPOutputStream;
-import org.apache.lucene.analysis.Analyzer;
-import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.codecs.lucene50.Lucene50StoredFieldsFormat;
 import org.apache.lucene.codecs.lucene70.Lucene70Codec;
 import org.apache.lucene.document.DateTools;
@@ -83,6 +81,7 @@ import org.apache.lucene.store.LockFactory;
 import org.apache.lucene.store.NativeFSLockFactory;
 import org.apache.lucene.store.NoLockFactory;
 import org.apache.lucene.store.SimpleFSLockFactory;
+import org.apache.lucene.util.Bits;
 import org.apache.lucene.util.BytesRef;
 import org.opensolaris.opengrok.analysis.AnalyzerGuru;
 import org.opensolaris.opengrok.analysis.Ctags;
@@ -94,7 +93,6 @@ import org.opensolaris.opengrok.configuration.Project;
 import org.opensolaris.opengrok.configuration.RuntimeEnvironment;
 import org.opensolaris.opengrok.configuration.messages.Message;
 import org.opensolaris.opengrok.history.HistoryException;
-import org.opensolaris.opengrok.history.HistoryGuru;
 import org.opensolaris.opengrok.logger.LoggerFactory;
 import org.opensolaris.opengrok.search.QueryBuilder;
 import org.opensolaris.opengrok.util.ForbiddenSymlinkException;
@@ -131,6 +129,7 @@ public class IndexDatabase {
     private IndexWriter writer;
     private IndexAnalysisSettings2 settings;
     private PendingFileCompleter completer;
+    private Bits liveDocs;
     private TermsEnum uidIter;
     private PostingsEnum postsIter;
     private IgnoredNames ignoredNames;
@@ -402,6 +401,7 @@ public class IndexDatabase {
         settings = null;
         uidIter = null;
         postsIter = null;
+        liveDocs = null;
         acceptedNonlocalSymlinks.clear();
 
         IOException finishingException = null;
@@ -453,6 +453,7 @@ public class IndexDatabase {
 
                 String startuid = Util.path2uid(dir, "");
                 reader = DirectoryReader.open(indexDirectory); // open existing index
+                liveDocs = MultiFields.getLiveDocs(reader);
                 settings = readAnalysisSettings();
                 if (settings == null) {
                     settings = new IndexAnalysisSettings2();
@@ -492,7 +493,9 @@ public class IndexDatabase {
                     while (uidIter != null && uidIter.term() != null
                         && uidIter.term().utf8ToString().startsWith(startuid)) {
 
-                        removeFile(true);
+                        if (anyLiveDoc()) {
+                            removeFile(true);
+                        }
                         BytesRef next = uidIter.next();
                         if (next == null) {
                             uidIter=null;
@@ -533,104 +536,16 @@ public class IndexDatabase {
         if (finishingException != null) throw finishingException;
 
         if (!isInterrupted() && isDirty()) {
-            if (env.isOptimizeDatabase()) {
-                optimize();
-            }
-            env.setIndexTimestamp();
-        }
-    }
-
-    /**
-     * Optimize all index databases
-     *
-     * @param parallelizer a defined instance
-     * @throws IOException if an error occurs
-     */
-    static CountDownLatch optimizeAll(RuntimeEnvironment env)
-            throws IOException {
-        List<IndexDatabase> dbs = new ArrayList<>();
-        IndexerParallelizer parallelizer = env.getIndexerParallelizer();
-        if (env.hasProjects()) {
-            for (Project project : env.getProjectList()) {
-                dbs.add(new IndexDatabase(env, project));
-            }
-        } else {
-            dbs.add(new IndexDatabase(env));
-        }
-
-        CountDownLatch latch = new CountDownLatch(dbs.size());
-        for (IndexDatabase d : dbs) {
-            final IndexDatabase db = d;
-            if (db.isDirty()) {
-                parallelizer.getFixedExecutor().submit(new Runnable() {
-                    @Override
-                    public void run() {
-                        try {
-                            db.update();
-                        } catch (Throwable e) {
-                            LOGGER.log(Level.SEVERE,
-                                "Problem updating lucene index database: ", e);
-                        } finally {
-                            latch.countDown();
-                        }
-                    }
-                });
-            }
-        }
-        return latch;
-    }
-
-    /**
-     * Optimize the index database
-     */
-    public void optimize() throws IOException {
-        synchronized (lock) {
-            if (running) {
-                LOGGER.warning("Optimize terminated... Someone else is updating / optimizing it!");
-                return;
-            }
-            running = true;
-        }
-
-        IndexWriter wrt = null;
-        IOException writerException = null;
-        try {
-            Statistics elapsed = new Statistics();
-            String projectDetail = this.project != null ? " for project " + project.getName() : "";
-            LOGGER.log(Level.INFO, "Optimizing the index{0}", projectDetail);
-            Analyzer analyzer = new StandardAnalyzer();
-            IndexWriterConfig conf = new IndexWriterConfig(analyzer);
-            conf.setOpenMode(OpenMode.CREATE_OR_APPEND);
-
-            wrt = new IndexWriter(indexDirectory, conf);
-            wrt.forceMerge(1); // this is deprecated and not needed anymore
-            elapsed.report(LOGGER, String.format("Done optimizing index%s", projectDetail));
             synchronized (lock) {
                 if (dirtyFile.exists() && !dirtyFile.delete()) {
-                    LOGGER.log(Level.FINE, "Failed to remove \"dirty-file\": {0}",
+                    LOGGER.log(Level.FINE,
+                        "Failed to remove \"dirty-file\": {0}",
                         dirtyFile.getAbsolutePath());
                 }
                 dirty = false;
             }
-        } catch (IOException e) {
-            writerException = e;
-            LOGGER.log(Level.SEVERE, "ERROR: optimizing index: {0}", e);
-        } finally {
-            if (wrt != null) {
-                try {
-                    wrt.close();
-                } catch (IOException e) {
-                    if (writerException == null) writerException = e;
-                    LOGGER.log(Level.WARNING,
-                        "An error occurred while closing writer", e);
-                }
-            }
-            synchronized (lock) {
-                running = false;
-            }
+            env.setIndexTimestamp();
         }
-
-        if (writerException != null) throw writerException;
     }
 
     private boolean isDirty() {
@@ -1090,14 +1005,19 @@ public class IndexDatabase {
                                 && uidIter.term().compareTo(emptyBR) != 0
                                 && uidIter.term().compareTo(buid) < 0) {
 
-                            // If the term's path matches path of currently processed file,
-                            // it is clear that the file has been modified and thus
-                            // removeFile() will be followed by call to addFile() below.
-                            // In such case, instruct removeFile() not to remove history
-                            // cache for the file so that incremental history cache
-                            // generation works.
-                            String termPath = Util.uid2url(uidIter.term().utf8ToString());
-                            removeFile(!termPath.equals(path));
+                            if (anyLiveDoc()) {
+                                String termPath = Util.uid2url(uidIter.term().
+                                        utf8ToString());
+                                // If the term's path matches path of currently
+                                // processed file, it is clear that the file
+                                // has been modified and thus removeFile() will
+                                // be followed by call to addFile() below.  In
+                                // such case, instruct removeFile() not to
+                                // remove history cache for the file so that
+                                // incremental history cache
+                                // generation works.
+                                removeFile(!termPath.equals(path));
+                            }
 
                             BytesRef next = uidIter.next();
                             if (next == null) {
@@ -1665,8 +1585,18 @@ public class IndexDatabase {
         postsIter = uidIter.postings(postsIter);
         while (postsIter.nextDoc() != DocIdSetIterator.NO_MORE_DOCS) {
             ++n;
+            int docID = postsIter.docID();
+            // If the document is deleted, break to skip further checking.
+            if (liveDocs != null && (docID >= liveDocs.length() ||
+                    !liveDocs.get(docID))) {
+                if (LOGGER.isLoggable(Level.FINEST)) {
+                    LOGGER.log(Level.FINEST, "not live doc {0}: {1}",
+                            new Object[]{docID, path});
+                }
+                continue;
+            }
             // Read a limited-fields version of the document.
-            Document doc = reader.document(postsIter.docID(), CHECK_FIELDS);
+            Document doc = reader.document(docID, CHECK_FIELDS);
             if (doc == null) {
                 LOGGER.log(Level.FINER, "No Document: {0}", path);
                 continue;
@@ -1780,6 +1710,18 @@ public class IndexDatabase {
             LOGGER.log(Level.FINEST, "Missing {0}", path);
         }
         return chkres;
+    }
+
+    private boolean anyLiveDoc() throws IOException {
+        postsIter = uidIter.postings(postsIter);
+        while (postsIter.nextDoc() != DocIdSetIterator.NO_MORE_DOCS) {
+            int docID = postsIter.docID();
+            if (liveDocs == null || (docID < liveDocs.length() &&
+                    liveDocs.get(docID))) {
+                return true;
+            }
+        }
+        return false;
     }
 
     private class IndexDownArgs {
